Better URL parsing/matching.
Now storing URI fragments (host, path, query, fragment) directly on document model
so that we can more easily determine proper URL uniqueness and better url searching from the extension. This will let erroneous things like nonsense from query params or nonsense fragments not interfere with the ability to find a document by the url.

We can also extend the extension to know that it is on a *site* with reviewed document(s) since we can search by host, instead of only highlighting the extension when you're on an exact contract page.